refactor: CalulateItem프로토콜과 관련된 타입을 CalculateItem.swift로 분리
