+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -7,18 +7,6 @@
 
 import Foundation
 
-protocol CalculateItem {
-    
-}
-
-enum Operator: CalculateItem {
-    case plus, minus, multiply, divide, equal
-}
-
-struct Number: CalculateItem {
-    let num: Double
-}
-
 class CalculatorItemQueue {
     var queue: [CalculateItem] = []
